refactor: don't expect ows:Keyword to have a Type
The Petascope UI turns non-functional if the underlying service has a list of
`ows:Keyword`s and not all of them have an `ows:Type`. This disables the
access to the `ows:Type` child.

This has no consequences (other than the UI working properly with Keywords
present) as the `ows:Type` is not shown anywhere in the UI.

SVC-1161